requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
python-jose[cryptography]
passlib
email-validator
requests
paho-mqtt
